+++ Kernel/System/Ticket.pm
@@ -6116,6 +6116,13 @@ sub TicketMerge {
         UserID       => $Param{UserID},
     );
 
+    # Update change time and user ID for main ticket.
+    #   See bug#13092 for more information.
+    return if !$DBObject->Do(
+        SQL  => 'UPDATE ticket SET change_time = current_timestamp, change_by = ? WHERE id = ?',
+        Bind => [ \$Param{UserID}, \$Param{MainTicketID} ],
+    );
+
     # set new state of merge ticket
     $Self->StateSet(
         State    => 'merged',